Restructure Assessment & Fix TabsView spacing
